Backed out changeset ada34cba0052 (bug 1348941) for image-related bustages CLOSED TREE
UltraBlame original commit: 8778b9d432f1cbef670e460d52cd255e631dad04

+++ image/encoders/ico/nsICOEncoder.h
@@ -50,14 +50,14 @@ public:
 protected:
   ~nsICOEncoder();
 
-  nsresult ParseOptions(const nsAString& aOptions, uint16_t& aBppOut,
+  nsresult ParseOptions(const nsAString& aOptions, uint32_t& aBppOut,
                         bool& aUsePNGOut);
   void NotifyListener();
 
   
   void InitFileHeader();
   
-  void InitInfoHeader(uint16_t aBPP, uint8_t aWidth, uint8_t aHeight);
+  void InitInfoHeader(uint32_t aBPP, uint8_t aWidth, uint8_t aHeight);
   
   void EncodeFileHeader();
